fixed bugs, refactor CardPage to be functional component, use useParam inside CardPage
